Update Authelia config-map and helm-release.yaml
This commit updates the Authelia config-map.yaml and helm-release.yaml files. It changes the filename for the notifier filesystem from "/tmp/notification.txt" to "/hacks/notification.txt" in the config-map.yaml file. In the helm-release.yaml file, it updates the "tmp" section to "hacks" and adds a comment to stop using file notification.